clean up Degree Req sidebar components
